Consider crate binary if there's main.rs
This way, at least simplest binary crate flychecking works with modules.

Currently, flycheck works ok in src/main.rs as it considers the crate binary while you're in that file. When working in src/foo.rs, which is a module foo of same binary crate, current setup's crate type guessing logic breaks, as modules are not accounted for at all.

This makes crate type guessing work in "global" fashion: now flycheck-rust-crate-type is "bin" in every file that belongs to this crate, if src/main.rs exists.