Fix error related to multiple binary targets
When both 'src/lib.rs' and 'src/main.rs' are present in the project, we
default to invoking `cargo rustc`.  This command requires an explicit
binary name with the `--bin` flag, otherwise it causes the error
described in GH-23.

This commit adds logic to parse the output of `cargo read-manifest` to
determine the target type to check, based on the buffer file name.  If
the file name matches one of the targets, this target is chosen.
Otherwise, the first target is chosen as a default.

If the target is a binary, we also extract the binary name and set a new
`flycheck-rust-binary-name` variable (see flycheck/flycheck#958) to
solve the error.

I've tested this work with 'src/main.rs' binaries as well as multiples
binaries under 'src/bin'.

Closes GH-25, fixes GH-23, and adds logic to tackle GH-8.