Add function to explain the error under point
The rust compiler provides lengthy error explanations that would not be
practical to display at all times with flycheck.  However, we can
provide a convenience function to lookup the explanation for the error
under point.

This commit adds a function that calls `rustc --explain ID` with ID
being the code of the error under point.