Use cargo to find the build target
As flycheck/flycheck#1206 has added support for the full conventional project
layout, flycheck-rust should be able to automatically set the variables for
all files in the project as well.

This commit uses `cargo --read-manifest` to get a list of build targets for the
project, and try to determine the correct values for the -rust-crate-type and
-rust-binary-name variables.

As Cargo doesn't actually read the files to determine the targets, we have no
way to know if a file will be compiled, but if the project follows the
conventional layout, we should get it right most of the time.  If not, users can
always override the variables using local variables.

This commit adds test to make sure the association of files to targets works on
a dummy project.