Demote errors in `flycheck-rust-setup`
As discussed in [1], raising errors in `flycheck-rust-setup` in
combination with `global-flycheck-mode` will leave Emacs in an unusable
state.

The previous commit added a guard for catching errors when the cargo
executable wasn't found, but other errors could still happen (wrong
argument types, or through `json-read`).

To make sure no errors are raised, this commit wraps the body of
`flycheck-rust-setup` in `with-demoted-errors`.

[1]: https://github.com/flycheck/flycheck-rust/issues/40#issuecomment-253760883