add custom build support (#62)
Support build scripts

When we're editing 'build.rs', its target kind in `cargo metadata` is 'custom-build', which we can not use as build arguments.  However, that build script will be checked by `cargo check` if we build any target in the same package.

So, if the selected target is a 'custom-build' one, we just pick another target from the same package, if any.

This commit adds tests for the build script situation.